Waypoint 6 - Fixes to crash due to threading.. Move starnames into own file
Locks around the delete/add part of star names to prevent a crash if it
tried to delete while its using the array.  Delete is lazy, only does it
if it can
Starnames moved to own file to try and reduce size of formmap
lock around stargrid use of the displayed set, so display cannot swap
in, in functions called from threads.  Other functions have a
Debug.Assert check they are in the UI thread.
Will continue to test.
